Refactor, move logic to cache class
